# backend/api/utils/open-api/openapi.yaml
components:
  schemas:
    Body_Upload_Files_uploadFiles__post:
      properties:
        files:
          items:
            format: binary
            type: string
          title: Files
          type: array
      required:
      - files
      title: Body_Upload_Files_uploadFiles__post
      type: object
    FilesResponseModel:
      properties:
        urls:
          items:
            type: string
          title: Urls
          type: array
      title: FilesResponseModel
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    LinkMetaModel:
      properties:
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        image:
          anyOf:
          - type: string
          - type: 'null'
          title: Image
        topic:
          anyOf:
          - type: string
          - type: 'null'
          title: Topic
        url:
          title: Url
          type: string
      required:
      - url
      title: LinkMetaModel
      type: object
    Message:
      properties:
        image:
          anyOf:
          - type: string
          - type: 'null'
          title: Image
        link_meta:
          anyOf:
          - $ref: '#/components/schemas/LinkMetaModel'
          - type: 'null'
        text:
          title: Text
          type: string
      title: Message
      type: object
    MessageCreate:
      properties:
        content:
          $ref: '#/components/schemas/Message'
        thread_id:
          title: Thread Id
          type: string
      required:
      - content
      title: MessageCreate
      type: object
    MessageResponse:
      properties:
        content:
          $ref: '#/components/schemas/Message'
        created_at:
          format: date-time
          title: Created At
          type: string
        creator_id:
          title: Creator Id
          type: string
        id:
          title: Id
          type: string
        last_modified:
          format: date-time
          title: Last Modified
          type: string
        thread_id:
          title: Thread Id
          type: string
      required:
      - id
      - content
      - creator_id
      - created_at
      - last_modified
      - thread_id
      title: MessageResponse
      type: object
    MessagesResponse:
      properties:
        messages:
          items:
            $ref: '#/components/schemas/MessageResponse'
          title: Messages
          type: array
      required:
      - messages
      title: MessagesResponse
      type: object
    SessionInfo:
      properties:
        accessTokenPayload:
          title: Accesstokenpayload
          type: object
        email:
          title: Email
          type: string
        fullName:
          title: Fullname
          type: string
        last_login:
          title: Last Login
          type: string
        picture:
          title: Picture
          type: string
        sessionHandle:
          title: Sessionhandle
          type: string
        userId:
          title: Userid
          type: string
      required:
      - sessionHandle
      - userId
      - fullName
      - email
      - accessTokenPayload
      title: SessionInfo
      type: object
    Thread-Input:
      properties:
        headline:
          $ref: '#/components/schemas/Message'
        topic:
          title: Topic
          type: string
      required:
      - topic
      - headline
      title: Thread
      type: object
    Thread-Output:
      properties:
        headline:
          $ref: '#/components/schemas/Message'
        topic:
          title: Topic
          type: string
      required:
      - topic
      - headline
      title: Thread
      type: object
    ThreadCreate:
      properties:
        content:
          $ref: '#/components/schemas/Thread-Input'
      required:
      - content
      title: ThreadCreate
      type: object
    ThreadResponse:
      properties:
        content:
          $ref: '#/components/schemas/Thread-Output'
        created_at:
          format: date-time
          title: Created At
          type: string
        creator_id:
          title: Creator Id
          type: string
        id:
          title: Id
          type: string
        last_modified:
          format: date-time
          title: Last Modified
          type: string
      required:
      - id
      - content
      - creator_id
      - created_at
      - last_modified
      title: ThreadResponse
      type: object
    ThreadsResponse:
      properties:
        threads:
          items:
            $ref: '#/components/schemas/ThreadResponse'
          title: Threads
          type: array
      required:
      - threads
      title: ThreadsResponse
      type: object
    UserMap:
      properties:
        users:
          additionalProperties:
            $ref: '#/components/schemas/UserModel'
          title: Users
          type: object
      required:
      - users
      title: UserMap
      type: object
    UserModel:
      properties:
        _id:
          title: ' Id'
          type: string
        email:
          default: unknown email
          title: Email
          type: string
        last_login:
          default: unknown last login
          title: Last Login
          type: string
        name:
          default: unknown user
          title: Name
          type: string
        picture:
          default: unknown picture link
          title: Picture
          type: string
      required:
      - _id
      title: UserModel
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  title: FastAPI
  version: 0.1.0
openapi: 3.1.0
paths:
  /healthcheck:
    get:
      operationId: healthcheck_healthcheck_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      summary: Healthcheck
  /linkmeta:
    get:
      operationId: get_link_meta_linkmeta_get
      parameters:
      - in: query
        name: url
        required: true
        schema:
          title: Url
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Link Meta
  /message:
    post:
      description: The api will create a new message in a thread and return the newly
        created message
      operationId: create_message
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageCreate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
          description: Newly created message
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create a new message in a thread
      tags:
      - threads
  /messages:
    get:
      description: The api will return all messages in a thread
      operationId: get_messages
      parameters:
      - in: query
        name: thread_id
        required: true
        schema:
          title: Thread Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'
          description: List of messages in a thread
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get all messages in a thread
      tags:
      - threads
  /searchThreads:
    get:
      operationId: search_threads_searchThreads_get
      parameters:
      - in: query
        name: query
        required: true
        schema:
          title: Query
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadsResponse'
          description: Search threads by query and get matching threads
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Search Threads
      tags:
      - threads
  /searchTitles:
    get:
      operationId: search_titles
      parameters:
      - in: query
        name: query
        required: true
        schema:
          title: Query
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  type: string
                title: Response Search Titles
                type: array
          description: Search threads by query and get topic
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Search Titles
      tags:
      - threads
  /sessioninfo:
    get:
      operationId: secure_api_sessioninfo_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionInfo'
          description: Successful Response
      summary: Secure Api
      tags:
      - session
  /slack_user_token:
    post:
      operationId: slack_user_token_slack_user_token_post
      parameters:
      - in: query
        name: authcode
        required: true
        schema:
          title: Authcode
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Slack User Token
  /thread:
    post:
      description: The api will create a new thread and return the newly created thread
      operationId: create_thread
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThreadCreate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadResponse'
          description: Newly created thread
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create a new thread
      tags:
      - threads
  /threads:
    get:
      description: The api will return all threads
      operationId: get_threads
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadsResponse'
          description: List of threads
      summary: Get all threads
      tags:
      - threads
  /uploadFiles/:
    post:
      operationId: Upload_Files_uploadFiles__post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_Upload_Files_uploadFiles__post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilesResponseModel'
          description: Returns the uploaded file url
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Upload Files
      tags:
      - storage
      - storage
  /user:
    get:
      operationId: all_users_user_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMap'
          description: Get user information
      summary: All Users
      tags:
      - threads