# solid_frontend/src/api/models/Message.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { LinkMetaModel } from './LinkMetaModel';
export type Message = {
    image?: (string | null);
    link_meta?: (LinkMetaModel | null);
    text?: string;
};