# solid_frontend/src/api/models/MessageCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Message } from './Message';
export type MessageCreate = {
    content: Message;
    thread_id?: string;
};