# solid_frontend/src/api/models/MessageResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Message } from './Message';
export type MessageResponse = {
    content: Message;
    created_at: string;
    creator_id: string;
    id: string;
    last_modified: string;
    thread_id: string;
};