# solid_frontend/src/api/models/ThreadCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Thread_Input } from './Thread_Input';
export type ThreadCreate = {
    content: Thread_Input;
};