# solid_frontend/src/api/models/ThreadResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Thread_Output } from './Thread_Output';
export type ThreadResponse = {
    content: Thread_Output;
    created_at: string;
    creator_id: string;
    id: string;
    last_modified: string;
};